ground_station/pyproject.toml
[project]
name = "monitor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "api-lasertracker",
    "ipykernel>=7.2.0",
    "matplotlib>=3.10.8",
    "pandas>=3.0.1",
    "pyserial>=3.5",
    "scipy>=1.17.0",
]

[tool.uv.sources]
api-lasertracker = { path = "api_lasertracker-7.17.1.0-cp314-cp314-win_amd64.whl" }
